* Fixed some 64-bit compile warnings  * Fixed HTTP_POST operations  * Fixed a segfault building user metadata objects
git-svn-id: https://atmos-c.googlecode.com/svn/branches/atmos-c-2.0@50 cb61ccbc-2dfe-265f-a3f4-a677fceac7d0

+++ transport.h
@@ -40,7 +40,7 @@ typedef struct {
 #define MAX_HEADERS 1024
 
 typedef struct {
-    int return_code;
+    long return_code;
     char *response_body;
     size_t body_size;
     char *headers[MAX_HEADERS];
